Make all data types own their data.
I've decided that trying to be "zero-copy" (mostly to be "like nom") isn't
really benefiting me in any real sense, so I've dropped back to using
Strings everywhere rather than strs with explicit lifetimes.

This can be reverted if I have some revelation later and begin to understand
how all of this works.