Add IrcStream and IrcReader types which do I/O parts of this.
The IrcStream type is roughly analogous to TcpStream (which it wraps),
which allows for cloning and multiple ownership in the same way as
that type.

The IrcReader type presents an iterator-like view over the read side of
an IRC connection.

Also adds some changes to make parser errors a bit more transparent,
as errors are required to work a little better to use io::Result.